Move shared embed/completion methods to TraceVLABase (§3 batch 3)
Move _embed_prefix_with_images / _embed_action_suffix / _embed_trace_suffix / _completion_predict
from Pi0TraceVLA into TraceVLABase (add einops + trace_observation imports to the base; drop the
now-unused posemb_sincos/fourier_encode_2d from pi0_trace_vla.py). These methods use parameters but
create none, so the move is param-tree-neutral.

Verified: param-path fingerprint identical for all 8 trace/target configs (structural gate), and
the trace_vla full-FT smoke reproduces the baseline exactly (loss 2.4391, param_norm 2179.8127).
Committed atomically (base + the inheriting subclass).
